PRでいただいた指摘を修正

+++ myapp/db/schema.rb
@@ -20,8 +20,8 @@ ActiveRecord::Schema[7.0].define(version: 2023_02_03_070748) do
   end
 
   create_table "task_labels", charset: "utf8mb4", force: :cascade do |t|
-    t.bigint "task_id", null: false
-    t.bigint "label_id", null: false
+    t.integer "task_id", null: false
+    t.integer "label_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["label_id"], name: "index_task_labels_on_label_id"
